refactor: more outsorcing
