+++ cDatabase.py
@@ -1,6 +1,7 @@
-import asyncio
+import asyncio, csv, os, sqlite3
 from datetime import datetime, timedelta
-import sqlite3
+
+ROLES_CONFIG_FILE = "elo_roles.csv"
 
 
 def init_db():
@@ -191,3 +192,139 @@ async def clean_old_pending_matches():
         except Exception as e:
             print(f"Error cleaning pending matches: {e}")
         await asyncio.sleep(1800)
+
+
+async def generate_pairings(ctx, season_number):
+    """Generate pairings for the season"""
+    try:
+        conn = sqlite3.connect("elo_bot.db")
+        c = conn.cursor()
+
+        c.execute(
+            """SELECT id, elo
+                     FROM players
+                     WHERE signed_up = 1"""
+        )
+        players = c.fetchall()
+
+        if not players:
+            await ctx.send("❌ No players have signed up for the season!")
+            return False
+
+        role_ranges = []
+        if os.path.exists(ROLES_CONFIG_FILE):
+            with open(ROLES_CONFIG_FILE, mode="r") as csvfile:
+                reader = csv.DictReader(csvfile)
+                for row in reader:
+                    if (
+                        not row.get("role")
+                        or not row.get("min elo")
+                        or not row.get("max elo")
+                    ):
+                        continue
+                    try:
+                        role_ranges.append(
+                            {
+                                "name": row["role"].strip(),
+                                "min": int(row["min elo"]),
+                                "max": int(row["max elo"]),
+                            }
+                        )
+                    except ValueError:
+                        continue
+
+        role_ranges.sort(key=lambda x: x["min"], reverse=True)
+
+        groups = {}
+        for player_id, elo in players:
+
+            for role_range in role_ranges:
+                if role_range["min"] <= elo <= role_range["max"]:
+                    if role_range["name"] not in groups:
+                        groups[role_range["name"]] = []
+                    groups[role_range["name"]].append(player_id)
+                    break
+
+        if not groups:
+            await ctx.send("❌ Couldn't group players by league roles!")
+            return False
+
+        total_pairings = 0
+        for group_name, player_ids in groups.items():
+
+            subgroups = []
+            if len(player_ids) > 7:
+
+                import random
+
+                random.shuffle(player_ids)
+                subgroup_size = max(4, len(player_ids) // ((len(player_ids) // 7) + 1))
+                subgroups = [
+                    player_ids[i : i + subgroup_size]
+                    for i in range(0, len(player_ids), subgroup_size)
+                ]
+            else:
+                subgroups = [player_ids]
+
+            for i, subgroup in enumerate(subgroups):
+                subgroup_name = (
+                    group_name if len(subgroups) == 1 else f"{group_name}-{i + 1}"
+                )
+
+                from itertools import combinations
+
+                pairings = list(combinations(subgroup, 2))
+
+                for p1, p2 in pairings:
+
+                    c.execute(
+                        """INSERT INTO pairings
+                                     (player1_id, player2_id, season_number, group_name)
+                                 VALUES (?, ?, ?, ?)""",
+                        (p1, p2, season_number, subgroup_name),
+                    )
+
+                    total_pairings += 1
+
+        conn.commit()
+        await ctx.send(
+            f"✅ Generated {total_pairings} pairings for season {season_number}!"
+        )
+        return True
+
+    except Exception as e:
+        await ctx.send(f"❌ Error generating pairings: {e}")
+        return False
+    finally:
+        conn.close()
+
+
+def add_pending_rep(reporter_id, opponent_id, reporter_result):
+    conn = sqlite3.connect("elo_bot.db")
+    c = conn.cursor()
+    c.execute(
+        """INSERT INTO pending_reps (reporter_id, opponent_id, reporter_result)
+                 VALUES (?, ?, ?)""",
+        (reporter_id, opponent_id, reporter_result),
+    )
+    conn.commit()
+    conn.close()
+
+
+def get_pending_rep(reporter_id, opponent_id):
+    conn = sqlite3.connect("elo_bot.db")
+    c = conn.cursor()
+    cutoff_time = datetime.now() - timedelta(minutes=30)
+    cutoff_str = cutoff_time.strftime("%Y-%m-%d %H:%M:%S")
+    c.execute(
+        """SELECT *
+                 FROM pending_reps
+                 WHERE reporter_id = ?
+                   AND opponent_id = ?
+                   AND timestamp >= ?
+                 ORDER BY timestamp DESC LIMIT 1""",
+        (reporter_id, opponent_id, cutoff_str),
+    )
+    rep = c.fetchone()
+    conn.close()
+    return rep
